begginings of a fighter select system.
fixed a bug with the health bar on respawns. discovered a new bug with
health bar on fighters with different health values. -_-

sheilds will not be active unless you actually have sheild health.
otherwise blasters will directly hit the ship.

started to implement a system for choosing new fighters before
respawning. created a new fighter but only difference for that is in the
editor atm.